fix: 6주차 리뷰 부분 수정

+++ Leon/umc10th/src/main/java/com/example/umc10th/domain/review/dto/ReviewResponseDTO.java
@@ -3,13 +3,15 @@ package com.example.umc10th.domain.review.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 public class ReviewResponseDTO {
 
     @Getter
     @Builder
     public static class ReviewInfo {
         private Long reviewId;
-        private Integer rating;
+        private BigDecimal rating;
         private String content;
     }
 }
